21.11

+++ Praca-inżynierska/app/Http/Controllers/QueueController.php
@@ -8,6 +8,7 @@ use App\Providers\RouteServiceProvider;
 use App\Models\User;
 use App\Models\Queue;
 use App\Models\Charging_station;
+use App\Models\Charging_point;
 use Auth;
 use Carbon\Carbon;
 
@@ -15,9 +16,18 @@ class QueueController extends Controller
 {
     public function queueView(Charging_station $charging_station){
 
-        //dd(Queue::all()->where('charging_station_id', $charging_station->id)->first()->users->isEmpty());
+        if($charging_station->number_of_charging_points === 1){
 
-        //$queue_id = Queue::where('charging_station_id', $charging_station->id)->first()->id;
+            $charging_point = $charging_station->charging_points->first();
+            $charging_point2 = null;
+
+        }else if($charging_station->number_of_charging_points === 2){
+
+            $charging_point = $charging_station->charging_points->first();
+            $charging_point2 = $charging_station->charging_points->skip(1)->take(1)->first();
+
+        }
+
 
         if(Queue::all()->where('charging_station_id', $charging_station->id)->first() == null){
             $users = [];
@@ -25,7 +35,12 @@ class QueueController extends Controller
             $users = Queue::all()->where('charging_station_id', $charging_station->id)->first()->users;
         }
 
-        return view('queuePage', ['charging_station' => $charging_station, 'users' => $users]);
+        return view('queuePage', [
+            'charging_station' => $charging_station,
+            'users' => $users,
+            'charging_point' => $charging_point,
+            'charging_point2' => $charging_point2
+        ]);
     }
 
     public function enroll(Charging_station $charging_station, Request $request){
@@ -64,12 +79,10 @@ class QueueController extends Controller
         }
 
 
-        return back();
+        return back()->with('status', 'enrolled');
     }
 
     public function leave(Charging_station $charging_station){
-
-        $queue = Queue::all()->where('charging_station_id', $charging_station->id)->first();
 
         Auth::user()->update([
             'sign_up_time' => null,
@@ -79,6 +92,12 @@ class QueueController extends Controller
 
         Auth::user()->queue()->dissociate();//do testu ale niby działa
         Auth::user()->save();//do testu ale niby działa
+
+        $queue = Queue::all()->where('charging_station_id', $charging_station->id)->first();
+
+        if($queue == null){
+            return back();
+        }
 
         if($queue->users->isEmpty()){
             $queue->delete();
